backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.0
numpy
PyPDF2==3.0.1
pydantic==2.9.0
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
urllib3==2.2.1
